pyproject.toml
[project]
name = "adtech"
version = "0.1.0"
description = ""
authors = [
    {name = "Jorge Alvarez",email = "44485505+jorgehaq@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.12.11"
pytest-django = "^4.11.1"
mypy = "^1.17.1"

